// frontend/generated/models/ProfileFormDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProfileSkillDto } from './ProfileSkillDto';
import {
    ProfileSkillDtoFromJSON,
    ProfileSkillDtoFromJSONTyped,
    ProfileSkillDtoToJSON,
    ProfileSkillDtoToJSONTyped,
} from './ProfileSkillDto';

/**
 * 
 * @export
 * @interface ProfileFormDto
 */
export interface ProfileFormDto {
    /**
     * 
     * @type {string}
     * @memberof ProfileFormDto
     */
    overview?: string;
    /**
     * 
     * @type {string}
     * @memberof ProfileFormDto
     */
    education?: string;
    /**
     * 
     * @type {string}
     * @memberof ProfileFormDto
     */
    phoneNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof ProfileFormDto
     */
    language?: string;
    /**
     * 
     * @type {Array<ProfileSkillDto>}
     * @memberof ProfileFormDto
     */
    skills?: Array<ProfileSkillDto>;
}

/**
 * Check if a given object implements the ProfileFormDto interface.
 */
export function instanceOfProfileFormDto(value: object): value is ProfileFormDto {
    return true;
}

export function ProfileFormDtoFromJSON(json: any): ProfileFormDto {
    return ProfileFormDtoFromJSONTyped(json, false);
}

export function ProfileFormDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProfileFormDto {
    if (json == null) {
        return json;
    }
    return {
        
        'overview': json['overview'] == null ? undefined : json['overview'],
        'education': json['education'] == null ? undefined : json['education'],
        'phoneNumber': json['phoneNumber'] == null ? undefined : json['phoneNumber'],
        'language': json['language'] == null ? undefined : json['language'],
        'skills': json['skills'] == null ? undefined : ((json['skills'] as Array<any>).map(ProfileSkillDtoFromJSON)),
    };
}

export function ProfileFormDtoToJSON(json: any): ProfileFormDto {
    return ProfileFormDtoToJSONTyped(json, false);
}

export function ProfileFormDtoToJSONTyped(value?: ProfileFormDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'overview': value['overview'],
        'education': value['education'],
        'phoneNumber': value['phoneNumber'],
        'language': value['language'],
        'skills': value['skills'] == null ? undefined : ((value['skills'] as Array<any>).map(ProfileSkillDtoToJSON)),
    };
}