// frontend/generated/models/ProfileSkillDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SkillDto } from './SkillDto';
import {
    SkillDtoFromJSON,
    SkillDtoFromJSONTyped,
    SkillDtoToJSON,
    SkillDtoToJSONTyped,
} from './SkillDto';
import type { ProficiencyEnum } from './ProficiencyEnum';
import {
    ProficiencyEnumFromJSON,
    ProficiencyEnumFromJSONTyped,
    ProficiencyEnumToJSON,
    ProficiencyEnumToJSONTyped,
} from './ProficiencyEnum';

/**
 * 
 * @export
 * @interface ProfileSkillDto
 */
export interface ProfileSkillDto {
    /**
     * 
     * @type {number}
     * @memberof ProfileSkillDto
     */
    profileSkillId?: number;
    /**
     * 
     * @type {SkillDto}
     * @memberof ProfileSkillDto
     */
    skill?: SkillDto;
    /**
     * 
     * @type {ProficiencyEnum}
     * @memberof ProfileSkillDto
     */
    proficiency?: ProficiencyEnum;
}



/**
 * Check if a given object implements the ProfileSkillDto interface.
 */
export function instanceOfProfileSkillDto(value: object): value is ProfileSkillDto {
    return true;
}

export function ProfileSkillDtoFromJSON(json: any): ProfileSkillDto {
    return ProfileSkillDtoFromJSONTyped(json, false);
}

export function ProfileSkillDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProfileSkillDto {
    if (json == null) {
        return json;
    }
    return {
        
        'profileSkillId': json['profileSkillId'] == null ? undefined : json['profileSkillId'],
        'skill': json['skill'] == null ? undefined : SkillDtoFromJSON(json['skill']),
        'proficiency': json['proficiency'] == null ? undefined : ProficiencyEnumFromJSON(json['proficiency']),
    };
}

export function ProfileSkillDtoToJSON(json: any): ProfileSkillDto {
    return ProfileSkillDtoToJSONTyped(json, false);
}

export function ProfileSkillDtoToJSONTyped(value?: ProfileSkillDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'profileSkillId': value['profileSkillId'],
        'skill': SkillDtoToJSON(value['skill']),
        'proficiency': ProficiencyEnumToJSON(value['proficiency']),
    };
}